Move IAM traits to its own package
This change allows us to evolve the IAM traits independent of the rest
of the AWS traits. We shouldn't put *all* AWS traits into a single
package because it requires that the traits are all changed in unison.
