+++ settings.gradle.kts
@@ -1,4 +1,5 @@
 rootProject.name = "smithy"
+include(":smithy-aws-iam-traits")
 include(":smithy-aws-traits")
 include(":smithy-aws-apigateway-openapi")
 include(":smithy-aws-protocol-tests")
