+++ smithy-aws-iam-traits/README.md
@@ -1,0 +1,4 @@
+# Smithy AWS IAM traits
+
+See the [Smithy specification](https://awslabs.github.io/smithy/spec/)
+for details on how these traits are used.
