+++ smithy-aws-iam-traits/src/main/resources/META-INF/services/software.amazon.smithy.model.validation.Validator
@@ -1,0 +1,1 @@
+software.amazon.smithy.aws.iam.traits.ConditionKeysValidator
