+++ smithy-aws-iam-traits/src/main/resources/META-INF/smithy/aws.iam.json
