+++ smithy-aws-iam-traits/src/main/resources/META-INF/smithy/manifest
@@ -1,0 +1,1 @@
+aws.iam.json
