+++ smithy-aws-iam-traits/src/test/resources/software/amazon/smithy/aws/iam/traits/required-actions.smithy
