+++ smithy-aws-traits/src/main/resources/META-INF/services/software.amazon.smithy.model.validation.Validator
@@ -1,5 +1,4 @@
 software.amazon.smithy.aws.traits.ArnTemplateValidator
 software.amazon.smithy.aws.traits.SdkServiceIdValidator
-software.amazon.smithy.aws.traits.iam.ConditionKeysValidator
 software.amazon.smithy.aws.traits.apigateway.AuthorizersTraitValidator
 software.amazon.smithy.aws.traits.clientendpointdiscovery.ClientEndpointDiscoveryValidator
