+++ smithy-aws-traits/src/main/resources/META-INF/smithy/manifest
@@ -1,5 +1,4 @@
 aws.api.json
 aws.apigateway.json
 aws.auth.json
-aws.iam.json
 aws.protocols.json
